requirements.txt
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
faiss-cpu==1.7.4
transformers==4.33.2
torch==2.0.1
tiktoken==0.4.0
reportlab==4.0.4
markdown==3.4.4
python-dotenv==1.0.0
tqdm==4.65.0
chromadb==0.4.5
langchain==0.0.340
langchain-community==0.0.10
streamlit==1.28.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8